fix1

+++ WM_HCA_Individual_Project.py
@@ -406,7 +406,7 @@ def gender():
     filtered_data = filtered_data[filtered_data["Entity"] != "World"]
 
     # Aggregate the data by taking the average smoking percentages and population for each country
-    agg_data = filtered_data.groupby(["Entity"]).mean().reset_index()
+    agg_data = filtered_data.groupby(["Entity"]).astype(float).mean().reset_index()
 
     # Calculate the scaling factor for the size of the circles
     circle_scaling_factor = 0.00005
